Revert "Feature login and logout for admin"
