--- src/main/java/com/cyan/hotel/repositoryService/AdminService.java
@@ -1,10 +1,0 @@
-package com.cyan.hotel.repositoryService;
-
-import com.cyan.hotel.model.Admin;
-
-public interface AdminService {
-  Boolean existByUsername(String username);
-
-  Admin findByUsername(String username);
-
-}
