+++ src/main/resources/application.properties
@@ -6,6 +6,6 @@ spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/ezhotel
 spring.datasource.username=root
-spring.datasource.password=babanh12
+spring.datasource.password=
 spring.datasource.data=classpath:import.sql
 spring.messages.basename=userValidation
