+++ src/main/webapp/resources/main.css
@@ -9,4 +9,3 @@ img {
   max-width: 100%;
   max-height: 100%;
 }
-
